requirements.txt
openai==0.27.2
PyPDF2==3.0.1
python-dotenv==1.0.0